gaia_agent/requirements.txt
gradio
requests
pandas
python-dotenv
smolagents
duckduckgo-search
openpyxl
PyPDF2
